ДЗ 2: Некоторая путаница с модулями
