+++ homework-02/pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.otus</groupId>
-    <artifactId>homework_01</artifactId>
+    <artifactId>homework_02</artifactId>
     <version>1.0-SNAPSHOT</version>
 
     <properties>
