+++ homework-02/src/main/java/ru/otus/homework/service/InputOutputServiceImpl.java
@@ -1,28 +1,30 @@
 package ru.otus.homework.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.cglib.core.Local;
-import org.springframework.context.MessageSource;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.support.ReloadableResourceBundleMessageSource;
 import org.springframework.stereotype.Service;
 
-import java.util.Locale;
+import java.io.InputStream;
+import java.io.PrintStream;
 import java.util.Scanner;
 
 @Service
 public class InputOutputServiceImpl implements InputOutputService {
-    private Scanner scanner;
+    private static final InputStream IN_STREAM = System.in;
+    private static final PrintStream OUT_STREAM = System.out;
+
+    private final PrintStream outStream;
+    private final Scanner scanner;
 
     public InputOutputServiceImpl() {
-        this.scanner = new Scanner(System.in);
+        this.outStream = OUT_STREAM;
+        this.scanner = new Scanner(IN_STREAM);
     }
 
     public void showMessage(String message) {
-        System.out.println(message);
+        outStream.println(message);
     }
 
     public String getMessage() {
         return scanner.next();
     }
 }
+
