+++ homework-02/src/main/java/ru/otus/homework/service/LocalizationService.java
@@ -1,5 +1,5 @@
 package ru.otus.homework.service;
 
 public interface LocalizationService {
-    String localizeMessage(String bundleCode);
+    String localizeMessage(String bundleCode, Object[] var2);
 }
