+++ homework-02/src/main/resources/application.properties
@@ -1,2 +1,2 @@
-qna.filepath=QuestionsAndAnswers.csv
-locale=ru-RU
+qna.filepath=csv/QuestionsAndAnswers.csv
+language.tag=ru-RU
