+++ homework-02/src/main/resources/messages_ru_RU.properties
@@ -1,1 +1,4 @@
-test.name=\u0422\u0435\u0441\u0442
+test.name=Тест
+test.firstName=Введите ваше имя
+test.lastName=Введите вашу фамилию
+test.result={0}, ваш результат {1}/{2}
