--- homework_02/.gitignore
@@ -1,4 +1,0 @@
-# Default ignored files
-.idea/
-target/
-*.iml
