--- homework_02/Readme.md
@@ -1,12 +1,0 @@
-# ДЗ 2. Добавить файл настроек, Annotation- + Java-based конфигурация приложения
-Цель: конфигурировать Spring-приложения современным способом, как это и делается в современном мире.
-
-Результат: готовое современное приложение на чистом Spring.
-
-Описание задание:
-Выполняется на основе предыдущего домашнего задания.
-
-1. Переписать конфигурацию в виде Java + Annotation-based конфигурации.
-2. Локализовать выводимые сообщения и вопросы (в CSV-файле).
-3. Добавьте файл настроек для приложения тестирования студентов. В конфигурационный файл можно поместить путь до CSV-файла и/или текущую локаль, количество правильных ответов для зачёта - на Ваше усмотрение.
-4. Если Вы пишите интеграционные тесты, то не забудьте добавить аналогичный файл и для тестов.
