--- homework_02/src/main/java/ru/otus/homework/Main.java
@@ -1,28 +1,0 @@
-package ru.otus.homework;
-
-import org.springframework.context.MessageSource;
-import org.springframework.context.annotation.*;
-import org.springframework.context.support.ReloadableResourceBundleMessageSource;
-import ru.otus.homework.service.QuestionAndAnswerService;
-
-@PropertySource("classpath:application.properties")
-@Configuration
-@ComponentScan
-public class Main {
-
-    @Bean
-    public MessageSource messageSource() {
-        ReloadableResourceBundleMessageSource ms = new ReloadableResourceBundleMessageSource();
-        ms.setBasename("messages");
-        ms.setDefaultEncoding("UTF-8");
-        return ms;
-    }
-
-    public static void main(String[] args) {
-        AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext(Main.class);
-
-        QuestionAndAnswerService service = context.getBean(QuestionAndAnswerService.class);
-
-        service.startTest();
-    }
-}
