--- homework_02/src/main/java/ru/otus/homework/dao/QuestionAndAnswerDAO.java
@@ -1,10 +1,0 @@
-package ru.otus.homework.dao;
-
-import ru.otus.homework.domain.QuestionAndAnswer;
-import ru.otus.homework.exceptions.QuestionsLoadingException;
-
-import java.util.List;
-
-public interface QuestionAndAnswerDAO {
-    List<QuestionAndAnswer> questionAndAnswerList() throws QuestionsLoadingException;
-}
