--- homework_02/src/main/java/ru/otus/homework/domain/Person.java
@@ -1,19 +1,0 @@
-package ru.otus.homework.domain;
-
-public class Person {
-    private final String firstName;
-    private final String lastName;
-
-    public Person(String firstName, String lastName) {
-        this.firstName = firstName;
-        this.lastName = lastName;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-}
