--- homework_02/src/main/java/ru/otus/homework/domain/QuestionAndAnswer.java
@@ -1,19 +1,0 @@
-package ru.otus.homework.domain;
-
-public class QuestionAndAnswer {
-    private final String question;
-    private final String answer;
-
-    public QuestionAndAnswer(String Question, String Answer) {
-        this.question = Question;
-        this.answer = Answer;
-    }
-
-    public String getQuestion() {
-        return question;
-    }
-
-    public String getAnswer() {
-        return answer;
-    }
-}
