--- homework_02/src/main/java/ru/otus/homework/exceptions/QuestionsLoadingException.java
@@ -1,8 +1,0 @@
-package ru.otus.homework.exceptions;
-
-public class QuestionsLoadingException extends Exception {
-
-    public QuestionsLoadingException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
