--- homework_02/src/main/java/ru/otus/homework/service/InputOutputService.java
@@ -1,11 +1,0 @@
-package ru.otus.homework.service;
-
-/**
- * Сервис для работы с вводом/выводом
- */
-public interface InputOutputService {
-
-    void showMessage(String message);
-
-    String getMessage();
-}
