--- homework_02/src/main/java/ru/otus/homework/service/PersonService.java
@@ -1,7 +1,0 @@
-package ru.otus.homework.service;
-
-import ru.otus.homework.domain.Person;
-
-public interface PersonService {
-    Person getPerson();
-}
