--- homework_02/src/main/java/ru/otus/homework/service/QuestionAndAnswerService.java
@@ -1,5 +1,0 @@
-package ru.otus.homework.service;
-
-public interface QuestionAndAnswerService {
-    void startTest();
-}
