--- homework_02/src/main/resources/application.properties
@@ -1,2 +1,0 @@
-qna.filepath=csv/QuestionsAndAnswers.csv
-language.tag=ru-RU
