--- homework_02/src/main/resources/csv/QuestionsAndAnswers_en_US.csv
@@ -1,5 +1,0 @@
-1. How many days in January?,31
-2. How many eyes does a bird have?,2
-3. How many paws does a cat have?,4
-4. What number of this question?,4
-5. How many days in a non-leap year,365
