--- homework_02/src/main/resources/messages.properties
@@ -1,4 +1,0 @@
-test.name=Test
-test.firstName=Enter your first name
-test.lastName=Enter your last name
-test.result={0}, your result is {1}/{2}
