--- homework_02/src/main/resources/messages_ru_RU.properties
@@ -1,4 +1,0 @@
-test.name=Тест
-test.firstName=Введите ваше имя
-test.lastName=Введите вашу фамилию
-test.result={0}, ваш результат {1}/{2}
